Migrate code to processor directory + code overhaul
